+++ modules/disable-widgets.php
@@ -3,18 +3,28 @@
 namespace Ensoul\Rankz\DisableWidgets;
 
 /**
+ * Disable widgets
  *
  * You can enable/disable this feature in functions.php (or lib/setup.php if you're using Sage):
  * add_theme_support('rankz-disable-widgets');
  */
 
 /**
- * Remove widgets page from admin menu
+ * Disable widgets from admin menu
  */
-function remove_widgets_admin_menu() {
+function disable_widgets_admin_menu() {
   remove_submenu_page('themes.php', 'widgets.php');
 }
-add_action('admin_menu', __NAMESPACE__ . '\\remove_widgets_admin_menu');
+add_action('admin_menu', __NAMESPACE__ . '\\disable_widgets_admin_menu');
+
+/**
+ * Disable widgets from admin bar
+ */
+function disable_widgets_admin_bar() {
+  global $wp_admin_bar;
+  $wp_admin_bar->remove_menu('widgets');
+}
+add_action('wp_before_admin_bar_render', __NAMESPACE__ . '\\disable_widgets_admin_bar');
 
 /**
  * Redirect any user trying to access widgets page
@@ -27,12 +37,3 @@ function widgets_admin_menu_redirect() {
   }
 }
 add_action('admin_init', __NAMESPACE__ . '\\widgets_admin_menu_redirect');
-
-/**
- * Remove widgets from admin bar
- */
-function remove_widgets_admin_bar() {
-  global $wp_admin_bar;
-  $wp_admin_bar->remove_menu('widgets');
-}
-add_action('wp_before_admin_bar_render', __NAMESPACE__ . '\\remove_widgets_admin_bar');