+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename `remove-update-notive` to `disable-update-notice`
 * Rename `disable-customization` to `disable-customizer`
 * Improve `admin-login` module with more customization
+* Improve some modules functions
 
 ### 2.1.0: Gen 9th, 2018
 * Update `itsensoul` organization name with `helloensoul`